Trabalho2/requirements.txt
pyserial
serial
rpi.bme280
smbus2